Fix Gradle junit-platform-launcher ClassLoader issue
